[SpellData] A row's defense type, mechanic and power bar carry their enum types
Spell.DefenseType is a core.DefenseType, whose order is the client's, so DefenseTypeCore is a
plain read. Spell.Mechanic and Effect.Mechanic are dbcenums.Mechanic, Power.Type is
dbcenums.PowerType, and Power and PowerCost take one. The generator's rage check names
dbcenums.POWER_RAGE instead of a local constant. The generated rows are bare integers and
compile unchanged.

+++ sim/core/spelldata/resolve_spell_test.go
@@ -85,7 +85,7 @@ func resolverRows() []Spell {
 		},
 		{
 			ID: 800, Name: "Bleed", School: 1, DefenseType: 2, SpellLevel: 50, DurationMs: 21000,
-			Mechanic:              uint8(dbcenums.MECHANIC_BLEED),
+			Mechanic:              dbcenums.MECHANIC_BLEED,
 			GCDMs:                 1500,
 			StartRecoveryCategory: 133,
 			MaxStack:              5,
